165067148-fix(model): model file structure
- modified .sequelizerc
- moved model files

[Fixes #165067148]
