+++ .sequelizerc
@@ -2,7 +2,7 @@ const path = require('path');
 
 module.exports = {
   "config": path.resolve('./src/config', 'databaseConfig.js'),
-  "models-path": path.resolve('./src/models'),
+  "models-path": path.resolve('./src/db_api/models'),
   "seeders-path": path.resolve('./src/seeders'),
   "migrations-path": path.resolve('./src/migrations')
 };
