--- src/db_api/models/testtable.js
@@ -1,11 +1,0 @@
-export default (sequelize, DataTypes) => {
-  const TestTable = sequelize.define(
-    'TestTable',
-    {
-      title: DataTypes.STRING,
-      id: DataTypes.STRING,
-    },
-    {}
-  );
-  return TestTable;
-};
